+++ CHANGELOG.rst
@@ -27,6 +27,8 @@ Fixed
   <https://github.com/omni-us/jsonargparse/pull/662>`__).
 - Pydantic discriminated unions handled incorrectly (`#667
   <https://github.com/omni-us/jsonargparse/pull/667>`__).
+- Failure when a link target has an undefined parent (`#668
+  <https://github.com/omni-us/jsonargparse/pull/668>`__)
 
 
 v4.36.0 (2025-01-17)